Move `instamatic.collections` to `instamatic._collections`
